+++ CHANGES.md
@@ -1,3 +1,9 @@
+# v1 - 10/23/2024
+## Added
+- Added 26 art images
+- Added function in `global_artwork.js` for adding artwork
+## Removed
+- Removed 2 test imgs
 # v1h - 1 Hour Version - 10/23/2024
 ## Added
 - Added Art div for artwork